Move read trimmer to trimming module
