+++ src/pgmap/trimming/read_trimmer.py
